Implement built in function 'cd' and 'exit'
All built in functino are put in builtin.c source file, and are invoked
by the uniform interface fucntion call_builtin(). The reason to do this
is to hide the concrete implementation of each builtin function, and
it could improve the expansibility of the program.

Since builtin function may affect the environment of running shell itself,
call_builtin() will be invoked if the command only has one subcommand.